requirements.txt
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
marshmallow==3.11.1
marshmallow-sqlalchemy==0.25.0
PyMySQL==1.0.2
python-dotenv==0.17.1
six==1.16.0
SQLAlchemy==1.4.14
Werkzeug==1.0.1
